requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
Markdown==3.4.3
pipreqs==0.4.13
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==2.0.2
yarg==0.1.9
